+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Version changes are pinned to SDK releases.
 
 - Add skip to market refreshing if asset is halted. ([#304](https://github.com/zetamarkets/sdk/pull/304))
 - Add usdc account check and creation on withdraw. ([#303](https://github.com/zetamarkets/sdk/pull/303))
+- New override in getLeverage() to use executionInfo price as markPrice. ([#302](https://github.com/zetamarkets/sdk/pull/302))
 
 ## [1.12.2]